requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.22.0
bitsandbytes>=0.39.0
coqui-ai-tts>=1.0.0
sounddevice>=0.4.7
